Fix Gatsby breaking when deleting helpers.md
This is a problem within GraphQL then strikes, when queried properties don't exist. There is an open issue for this on Gatsby, where they are actively looking for a solution for this (issue #3344). Until then we have two options:

A) We always make sure, that all queried properties are used for the assigned template, where the query is running
B) We create our own node fields, using `setFieldsOnGraphQLNodeType`